Add default type for generic `CompT`
Release v1.7.9 added a new generic parameter to the traverse function. This works fine when generics are inferred (`traverse(node, visitors, state)`), but it breaks when explicit generic parameters are provided (`traverse<MyNode, MyState>(node, visitors, state)`) because the new parameter is missing.

This PR adds a default type to the new generic parameter, so that existing usage of the `traverse` function will still work.